Okay, it seems the service and adaptor view actualy opdates okay. But for some reason it does not persist when saving the price in edit. Maybe if i fix that it will update the price difference correctly aswell.
Another thing is i need to change the initial setup a bit. I need to make an initial request so i can populate the database with a top 10 list.
And then next time it runs the loop it wont fail in the async handler where it then matches 0 stocks because the database is empty.